requirements.txt
numpy~=1.19.5
torch~=1.9.0
pytorch-lightning~=1.3.6
nltk~=3.2.5
transformers~=4.7.0
h5py~=3.2.1
wandb~=0.10.32
sparsemax~=0.1.9
torchvision~=0.10.0
pillow~=7.1.2
entmax~=1.0